Critical bugfix for dynamic collision notifications
This is a critical bugfix for the dynamic collision notification system.
All the refactoring performed in the previous push introduced a bug
where an empty "voxel collision profile" would be passed for meshes that
used a common cache identifier across multiple instances of the mesh.
This fix is simple, we ensure that the collision profile is populated
onto "CollisionData". This code branch is executed for meshes that
already have a solution cached and available for their use. We just had
to populate the solution into the task by fetching it from the cache.