Release v1.3, support for unbound worlds!
This is a major release for supporting unbounds worlds. By using the new
manager DonNavigationManagerUnbound plugin users are no longer
restricted to navigation within the navigable bounds (which is currently
irrelevant for anyone using the unbound manager). Navigation queries can
be run from anywhere now.

Unbound navigation will obviously cost more performance than navigation
with bounds. Also the code for unbound navigation is somewhat rough and
there's a lot of code duplication that comes with this. Unbound
navigation is to be considered experimental for now.